Added column users.uac_disabled; Refactoring
Methods add_user and update_user now takes a hash of attributes to update/insert

id:530
Ticket: none
Release-Note: none

+++ lib/Foswiki/LoginManager/UnifiedLogin.pm
@@ -342,40 +342,43 @@ sub processProviderLogin {
     }
 
     if (ref($loginResult) eq 'HASH' && $loginResult->{cuid}) {
-        $this->userLoggedIn($loginResult->{cuid});
-        $session->logger->log(
-            {
-                level    => 'info',
-                action   => 'login',
-                webTopic => $web . '.' . $topic,
-                extra    => "AUTHENTICATION SUCCESS - $loginResult->{cuid} - "
-            }
-        );
-        $this->{_cgisession}->param( 'VALIDATION', encode_json($loginResult->{data} || {}) )
-          if $this->{_cgisession};
-        my ( $origurl, $origmethod, $origaction ) = _unpackRequest($provider->origin);
-        if (!$origurl || $origaction eq 'login') {
-            $origurl = $session->getScriptUrl(0, 'view', $web, $topic);
-            $session->{request}->delete_all;
-        } else {
-            # Unpack params encoded in the origurl and restore them
-            # to the query. If they were left in the query string they
-            # would be lost if we redirect with passthrough.
-            # First extract the params, ignoring any trailing fragment.
-            if ( $origurl =~ s/\?([^#]*)// ) {
-                foreach my $pair ( split( /[&;]/, $1 ) ) {
-                    if ( $pair =~ /(.*?)=(.*)/ ) {
-                        $session->{request}->param( $1, TAINT($2) );
+        my $deactivated = $this->{uac}->{db}->selectrow_array("SELECT deactivated FROM users WHERE cuid=?", {}, $loginResult->{cuid});
+        unless ($deactivated) {
+            $this->userLoggedIn($loginResult->{cuid});
+            $session->logger->log(
+                {
+                    level    => 'info',
+                    action   => 'login',
+                    webTopic => $web . '.' . $topic,
+                    extra    => "AUTHENTICATION SUCCESS - $loginResult->{cuid} - "
+                }
+            );
+            $this->{_cgisession}->param( 'VALIDATION', encode_json($loginResult->{data} || {}) )
+              if $this->{_cgisession};
+            my ( $origurl, $origmethod, $origaction ) = _unpackRequest($provider->origin);
+            if (!$origurl || $origaction eq 'login') {
+                $origurl = $session->getScriptUrl(0, 'view', $web, $topic);
+                $session->{request}->delete_all;
+            } else {
+                # Unpack params encoded in the origurl and restore them
+                # to the query. If they were left in the query string they
+                # would be lost if we redirect with passthrough.
+                # First extract the params, ignoring any trailing fragment.
+                if ( $origurl =~ s/\?([^#]*)// ) {
+                    foreach my $pair ( split( /[&;]/, $1 ) ) {
+                        if ( $pair =~ /(.*?)=(.*)/ ) {
+                            $session->{request}->param( $1, TAINT($2) );
+                        }
                     }
                 }
-            }
 
-            # Restore the action too
-            $session->{request}->action($origaction) if $origaction;
+                # Restore the action too
+                $session->{request}->action($origaction) if $origaction;
+            }
+            $session->{request}->method($origmethod);
+            $session->redirect($origurl, 1);
+            return $loginResult->{cuid};
         }
-        $session->{request}->method($origmethod);
-        $session->redirect($origurl, 1);
-        return $loginResult->{cuid};
     }
 
     if ($Foswiki::cfg{UnifiedAuth}{DefaultAuthProvider}) {
